syntax and location of PROMPT_COMMAND fixed
This was a struggle, but I think this is finally the right syntax.  The PROMPT_COMMAND is set after the disk is mounted and the device has an IP.

Any commands entered are logged to /var/log/system.log.  Use Console to view them or cat /var/log/system.log | grep "SUM-IDS"